requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dateutil>=2.8.0

# Data processing and analysis
scipy>=1.9.0
scikit-learn>=1.2.0
statsmodels>=0.13.0
ta-lib>=0.4.25

# Database and caching
sqlalchemy>=1.4.0
redis>=4.3.0
pymongo>=4.3.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=0.19.0
loguru>=0.6.0

# Async and concurrency
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0

# API and web framework
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.10.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Code quality
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=1.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0

# Optional trading integrations
alpaca-py>=0.8.0
ccxt>=4.0.0
yfinance>=0.1.87

# Visualization
plotly>=5.10.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0